perf(sidebar): defer authenticated menus
